+++ l10n_ro_config/readme/DESCRIPTION.rst
@@ -5,4 +5,4 @@ For Romanian companies, you will have a dedicated page in Configuration, from wh
 
 \* All changes added by the Romanian localization will only apply for companies that have the "Use Romanian Accounting" checkbox set.
 
-\* All menus and fields are visible in the interface if you are working on a Romanian company, otherwise the are hidden.
+
