+++ l10n_ro_isolation/__init__.py
@@ -1,0 +1,1 @@
+from . import models
