+++ l10n_ro_isolation/__manifest__.py
@@ -3,25 +3,21 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 {
-    "name": "Romania - Localization Config",
-    "summary": "Romania - Localization Install and Config Applications",
+    "name": "Romania - isolation",
+    "summary": "Romania - isolation ",
     "license": "AGPL-3",
-    "version": "16.0.1.12.1",
+    "version": "16.0.1.0.0",
     "author": "NextERP Romania,"
     "Forest and Biomass Romania,"
     "Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/l10n-romania",
     "depends": ["l10n_ro", "base_vat"],
     "data": [
-        "views/account_journal.xml",
-        "views/common_report.xml",
+        "security/ro_menus_group.xml",
         "views/res_config_view.xml",
-        "views/res_bank_view.xml",
-        "views/res_partner_view.xml",
     ],
-    "pre_init_hook": "pre_init_hook",
     "development_status": "Mature",
-    "maintainers": ["feketemihai"],
+    "maintainers": ["dhongu", "feketemihai"],
     "installable": True,
     "auto_install": True,
 }
