+++ l10n_ro_isolation/models/__init__.py
@@ -1,0 +1,2 @@
+from . import l10n_ro_mixin
+from . import res_company
