+++ l10n_ro_isolation/readme/CONTRIBUTORS.rst
@@ -1,0 +1,11 @@
+* `NextERP Romania <https://www.nexterp.ro>`_:
+
+  * Fekete Mihai <feketemihai@nexterp.ro>
+
+* `Terrabit <https://www.terrabit.ro>`_:
+
+  * Dorin Hongu <dhongu@gmail.com>
+
+* Adrian Vasile <adrian.vasile@gmail.com>
+
+Do not contact contributors directly about support or help with technical issues.
