+++ l10n_ro_isolation/readme/DESCRIPTION.rst
@@ -1,0 +1,4 @@
+
+\* All changes added by the Romanian localization will only apply for companies that have the "Use Romanian Accounting" checkbox set.
+
+\* All menus and fields are visible in the interface if you are working on a Romanian company, otherwise the are hidden.
