+++ l10n_ro_isolation/readme/INSTALL.rst
@@ -3,5 +3,5 @@ To install this module, you need to:
 * clone the branch 16.0 of the repository https://github.com/OCA/l10n-romania
 * add the path to this repository in your configuration (addons-path)
 * update the module list
-* search for "Romania - Localization Config" in your addons
+* search for "Romania - isolation" in your addons
 * install the module
