+++ l10n_ro_isolation/views/res_config_view.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="res_config_settings_view_form" model="ir.ui.view">
+        <field name="name">res.config.settings.view.form.inherit</field>
+        <field name="model">res.config.settings</field>
+        <field name="priority" eval="50" />
+        <field name="inherit_id" ref="l10n_ro_config.res_config_settings_view_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//div[@data-key='l10n_ro_config']" position="attributes">
+                <attribute name="groups">l10n_ro_isolation.group_ro_menus</attribute>
+            </xpath>
+        </field>
+    </record>
+</odoo>
