+++ setup/l10n_ro_isolation/odoo/addons/l10n_ro_isolation
@@ -1,0 +1,1 @@
+../../../../l10n_ro_isolation
